Add compact int/long tests.
